Refactor(auth): Implements user authentication
Refactors API client and onboarding logic.
Improves authentication flow using Google Sign-In.
Handles user onboarding status.
Renames clients to controllers for clarity.

Related to: 3-handle-response-codes-in-apiclient
